Minor installation script fixes
The initial "apt-get install" invocation was missing "sudo" and hence
those packages would fail to install if the script were run as an
unprivileged user (as is implied by the use of "sudo" elsewhere).

Also, some options were added to the "apt-get upgrade" statement to
ensure that the upgrade is done in unattended mode and the system will
not prompt for answers if necessary.