Bootstrap script fixes and extensions
The "sudo apt-get upgrade && sudo apt-get update" commands are in the
wrong order.  The package lists should be updated (via "apt-get
update") before the packages are upgraded (via "apt-get upgrade").  As
things stand, the "sudo apt-get upgrade" will upgrade zero packages
because they will all be up to date already (because the package lists
reference when the VM image was created).

This fix simply updates the package lists and doesn't upgrade the rest
of the packages.  This is so that the installation of Git will succeed.
 Instead, we postpone updating of all the packages to when it is done
as part of "utils/client_setup.sh".

We also change ubuntu_client_bootstrap.sh so that the RABBITMQ_URL can
optionally be passed in as a parameter when invoking the script.  This
paves the way for provisioning via, e.g., the Vagrant shell provisioner
and passing in the RABBITMQ_URL value when setting up the machine that
way.